Reorganize and consolidate records related features into one folder
