--- backend/src/RecordTracker.API/Features/RecordFields/Dtos/RecordFieldDto.cs
@@ -1,12 +1,0 @@
-﻿using RecordTracker.Infrastructure.Entities;
-
-namespace RecordTracker.API.Features.RecordFields.Dtos;
-
-public class RecordFieldDto
-{
-    public Guid Id { get; set; }
-    public string Name { get; set; } = default!;
-    public FieldType FieldType { get; set; }
-    public bool IsRequired { get; set; }
-    public int Order { get; set; }
-}
