--- backend/src/RecordTracker.API/Features/RecordItems/Dtos/RecordItemDto.cs
@@ -1,6 +1,0 @@
-﻿namespace RecordTracker.API.Features.RecordItems.Dtos
-{
-    public class RecordItemDto
-    {
-    }
-}
