+++ backend/src/RecordTracker.API/Features/Records/Dtos/RecordFieldDto.cs
@@ -1,9 +1,10 @@
 ﻿using RecordTracker.Infrastructure.Entities;
 
-namespace RecordTracker.API.Features.RecordFields.Dtos;
+namespace RecordTracker.API.Features.Records.Dtos;
 
-public class CreateRecordFieldDto
+public class RecordFieldDto
 {
+    public Guid Id { get; set; }
     public string Name { get; set; } = default!;
     public FieldType FieldType { get; set; }
     public bool IsRequired { get; set; }
