+++ backend/src/RecordTracker.API/Features/Records/Dtos/RecordItemDto.cs
@@ -1,0 +1,6 @@
+﻿namespace RecordTracker.API.Features.Records.Dtos
+{
+    public class RecordItemDto
+    {
+    }
+}
